Improve platform module

+++ xxw-shop-applications/xxw-shop-service/xxw-shop-service-platform/src/main/java/com/xxw/shop/module/user/entity/SysUser.java
@@ -2,13 +2,13 @@ package com.xxw.shop.module.user.entity;
 
 import com.mybatisflex.annotation.Id;
 import com.mybatisflex.annotation.Table;
-import java.io.Serializable;
-import java.math.BigInteger;
-import java.time.LocalDateTime;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+
+import java.io.Serializable;
+import java.time.LocalDateTime;
 
 /**
  *  实体类。
@@ -26,7 +26,7 @@ public class SysUser implements Serializable {
      * 平台用户id
      */
     @Id
-    private BigInteger sysUserId;
+    private Long sysUserId;
     /**
      * 创建时间
      */
